split out scheduler from backend and fix circular imports
