+++ backend/model/strawberry/mutation_breakout/upsert_wikibase_mutation.py
@@ -8,10 +8,10 @@ from fetch_data import update_cloud_instances
 from model.enum import WikibaseType, WikibaseURLType
 from model.strawberry.input import WikibaseInput
 from model.strawberry.output import WikibaseStrawberryModel
-from resolvers import (
-    add_wikibase,
+from resolvers.add import add_wikibase
+from resolvers.authentication import authenticate
+from resolvers.update import (
     add_wikibase_language,
-    authenticate,
     remove_wikibase_language,
     remove_wikibase_url,
     update_wikibase_primary_language,
