--- backend/schedule/__init__.py
@@ -1,3 +1,0 @@
-"""Scheduler"""
-
-from schedule.scheduler import scheduler
